Implemented action and trigger cards.
The proposed action card and trigger card to update traveling details to
work address have been implemented.
In order to make sure the context of the tokens is explicit for the end
user, only a single work address is supported. Multiple addresses can be
supported in the future. Both the trigger tokens as general tokens are
updated so that users can create their own voice messages.